feat(unum1): implement native type conversions (#573)
* feat(unum1): implement native type conversions for unum Type I

Implement operator=(double) that encodes IEEE-754 doubles into unum
Type I format by decomposing into sign, exponent, and fraction, then
finding the tightest esize/fsize representation. The ubit is set when
the value cannot be represented exactly.

Key design decisions:
- IEEE-754-style subnormal convention: biased_exp=0 means hidden bit
  is 0 (subnormal/zero), biased_exp>0 means hidden bit is 1 (normal)
- Encoder finds smallest esize where biased_exp >= 1 (avoids
  subnormal zone for normal values)
- Encoder finds smallest fsize that exactly captures the fraction;
  sets ubit=1 if no exact match within max fsize
- to_double() updated to handle subnormal decoding

Also implements:
- Integer assignment via double conversion
- Unary negation (sign bit flip)
- Conversion test suite covering exact round-trips (powers of 2,
  binary fractions), inexact ubit flagging (1/3, pi), negative
  values, integer assignment, and unary negation

Resolves #565


* fix(unum1): update education example to use unsigned template params

The unum template changed from size_t to unsigned parameters in Phase 1.
Update basic_operators.cpp to match.


---------

+++ education/number/unum/basic_operators.cpp
@@ -7,7 +7,7 @@
 #include <universal/number/unum/unum.hpp>
 
 // quick helper to report on a posit's specialness
-template<size_t nbits, size_t es>
+template<unsigned nbits, unsigned es>
 void checkSpecialCases(sw::universal::unum<nbits, es> p) {
 	std::cout << "unum is " << (p.iszero() ? "zero " : "non-zero ") << (p.ispos() ? "positive " : "negative ") << (p.isnan() ? "Not a Number" : "Its a Real") << std::endl;
 }
@@ -17,8 +17,8 @@ int main()
 try {
 	using namespace sw::universal;	// standard namespace for unum
 
-	const size_t nbits = 16;
-	const size_t es = 5;
+	const unsigned nbits = 16;
+	const unsigned es = 5;
 	unum<nbits, es> p1, p2, p3, p4, p5, p6;
 
 //	/* constexpr */ double minpos = minpos_value<nbits, es>();
